+++ common/src/main/java/com/st0x0ef/stellaris/common/menus/PumpjackMenu.java
@@ -51,12 +51,12 @@ public class PumpjackMenu extends BaseContainer {
         if (!player.level().isClientSide()) {
 
             NetworkManager.sendToPlayer(player, new SyncWidgetsTanksPacket(
-                    new long[] {blockEntity.getResultTank().getAmount(), blockEntity.chunkOilLevel()},
+                    new int[] {blockEntity.getResultTank().getAmount(), blockEntity.chunkOilLevel()},
                     new ResourceLocation[] {blockEntity.getResultTank().getStack().getFluid().arch$registryName()}
             ));
 
             NetworkManager.sendToPlayer(player, new SyncWidgetsTanksPacket(
-                    new long[] {blockEntity.getEnergy(null).getEnergy()}
+                    new int[] {blockEntity.getEnergy(null).getEnergy()}
             ));
 
         }
