+++ common/src/main/java/com/st0x0ef/stellaris/common/menus/WaterSeparatorMenu.java
@@ -60,12 +60,12 @@ public class WaterSeparatorMenu extends BaseContainer {
             ));
 
             NetworkManager.sendToPlayer(player, new SyncWidgetsTanksPacket(
-                    new long[] {blockEntity.ingredientTank.getAmount()},
+                    new int[] {blockEntity.ingredientTank.getAmount()},
                     new ResourceLocation[] {blockEntity.ingredientTank.getStack().getFluid().arch$registryName()}
             ));
 
             NetworkManager.sendToPlayer(player, new SyncWidgetsTanksPacket(
-                    new long[] {blockEntity.getEnergy(null).getEnergy(), 0, 0}
+                    new int[] {blockEntity.getEnergy(null).getEnergy(), 0, 0}
             ));
         }
     }
